requirements.txt
fontTools

GitPython==3.1.0
PyQt5>=5.14
psutil==5.7.0
pynput>=1.6
pyshortcuts==1.7
pystray==0.14.4
setuptools>=45.0.0
